AqXCo 0.55
Incluido:
-colisionComponent
-escenarioComponent
-rotarSystem
Modificado:
-facadeMotorGrafico
-camaraSystem
-moveSystem
-main

+++ DESARROLLO/HITO1/ITERACION 3/ArquitecturaXComponentes/entitySystem/systems/camaraSystem.h
@@ -23,7 +23,7 @@ public:
     camaraSystem(const camaraSystem& orig);
     camaraSystem(entityManager*);
     virtual ~camaraSystem();
-    void update(unsigned int, unsigned int);
+    void update(double, unsigned int);
     void actualizarCamara(unsigned int, char, double);
 private:
